Regulatory annotation panel: update cursor shape
For the regulatory annotation panel component,
to make it easier to identify the the area (i.e. the ruler)
that can be used to initiate selection of a subregion,
the cursor changes its shape to a crosshair when it is positioned over it.

After subregion selection has started, the cursor retainst its crosshair shape
regardless of which element within the annotation panel it is positioned over,
until the subregion has been selected.