Structural variants: Include both reference and alt genomes in variant requests (#10)
Given that we will have multiple variants files where the reference genome is the same, but the other genome is different
(e.g. the reference genome against different haplotypes), it appears that need to include both the reference genome id
and the alt genome id in the url for the server to correctly identify the data file.